refactor: pull some builds into own files
