+++ lib/ci/ci_result_parser.js
@@ -1,9 +1,6 @@
 'use strict';
 
-const {
-  parsePRFromURL
-} = require('../links');
-const Cache = require('../cache');
+const { parsePRFromURL } = require('../links');
 const CIFailureParser = require('./ci_failure_parser');
 const {
   FAILURE_TYPES: {
@@ -17,7 +14,6 @@ const {
     [NCU_FAILURE]: NCUFailure,
     [RESUME_FAILURE]: ResumeFailure
   },
-  CIResult,
   FAILURE_TYPES_NAME
 } = CIFailureParser;
 const {
@@ -26,335 +22,25 @@ const {
   CI_TYPES
 } = require('./ci_type_parser');
 const {
-  flatten,
-  shortSha
-} = require('../utils');
+  BUILD_TREE,
+  COMMIT_TREE,
+  FANNED_TREE,
+  LINTER_TREE,
+  PR_TREE,
+  RUN_TREE
+} = require('./jenkins_constants');
+const { fold, statusType, pad } = require('./ci_utils');
+
+const { Job, jobCache } = require('./build-types/job');
+const { TestBuild } = require('./build-types/test_build');
+const { CITGMBuild } = require('./build-types/citgm_build');
+const { flatten } = require('../utils');
 const qs = require('querystring');
 const _ = require('lodash');
 const chalk = require('chalk');
 
-const SUCCESS = 'SUCCESS';
-const FAILURE = 'FAILURE';
-const ABORTED = 'ABORTED';
-const UNSTABLE = 'UNSTABLE';
-
-// com.tikal.jenkins.plugins.multijob.MultiJobBuild
-const BUILD_FIELDS = 'builtOn,buildNumber,jobName,result,url';
-const ACTION_TREE = 'actions[parameters[name,value]]';
-const CHANGE_FIELDS = 'commitId,author[absoluteUrl,fullName],authorEmail,' +
-                      'msg,date';
-const CHANGE_TREE = `changeSet[items[${CHANGE_FIELDS}]]`;
-const PR_TREE =
-  `result,url,number,${ACTION_TREE},${CHANGE_TREE},builtOn,` +
-  `subBuilds[${BUILD_FIELDS},build[subBuilds[${BUILD_FIELDS}]]]`;
-const COMMIT_TREE =
-  `result,url,number,${ACTION_TREE},${CHANGE_TREE},builtOn,` +
-  `subBuilds[${BUILD_FIELDS}]`;
-const CITGM_MAIN_TREE =
-  `result,url,number,${ACTION_TREE},${CHANGE_TREE},builtOn`;
-
-// com.tikal.jenkins.plugins.multijob.MultiJobBuild
-const FANNED_TREE =
-  `result,url,number,subBuilds[phaseName,${BUILD_FIELDS}]`;
-
-// hudson.matrix.MatrixBuild
-const BUILD_TREE = 'result,runs[url,number,result],builtOn';
-const LINTER_TREE = 'result,url,number,builtOn';
-const CAUSE_TREE = 'upstreamBuild,upstreamProject,shortDescription,_class';
-const RUN_TREE = `actions[causes[${CAUSE_TREE}]],builtOn`;
-
-// hudson.tasks.test.MatrixTestResult
-const RESULT_TREE = 'result[suites[cases[name,status]]]';
-const CITGM_REPORT_TREE =
-  `failCount,skipCount,totalCount,childReports[child[url],${RESULT_TREE}]`;
-
 function getPath(url) {
   return url.replace(`https://${CI_DOMAIN}/`, '').replace('api/json', '');
-}
-
-function getUrl(path) {
-  return `https://${CI_DOMAIN}/${getPath(path)}`;
-}
-
-function resultName(result) {
-  return result === null ? 'PENDING' : result.toUpperCase();
-}
-
-function fold(summary, code) {
-  const dataBlock = '```\n' + code + '\n```';
-  const summaryBlock = `\n<summary>${summary}</summary>\n`;
-  return `<details>${summaryBlock}\n${dataBlock}\n</details>`;
-}
-
-function getNodeName(url) {
-  const re = /\/nodes=(.+?)\//;
-  if (re.test(url)) {
-    return url.match(re)[1];
-  }
-  const parts = url.split('/');
-  return parts[parts.length - 3];
-}
-
-class Job {
-  constructor(cli, request, path, tree) {
-    this.cli = cli;
-    this.request = request;
-    this.path = path;
-    this.tree = tree;
-  }
-
-  get jobUrl() {
-    const { path } = this;
-    return `https://${CI_DOMAIN}/${path}`;
-  }
-
-  get apiUrl() {
-    const { tree } = this;
-    const query = tree ? `?tree=${qs.escape(tree)}` : '';
-    return `${this.jobUrl}api/json${query}`;
-  }
-
-  get consoleUrl() {
-    const { path } = this;
-    return `https://${CI_DOMAIN}/${path}consoleText`;
-  }
-
-  get consoleUIUrl() {
-    const { path } = this;
-    return `https://${CI_DOMAIN}/${path}console`;
-  }
-
-  async getBuildData(type = 'Build') {
-    const { cli, path } = this;
-    cli.startSpinner(`Querying data for ${path}`);
-    const data = await this.getAPIData();
-    cli.stopSpinner(`${type} data downloaded`);
-    return data;
-  }
-
-  getCause(actions) {
-    if (actions && actions.find(item => item.causes)) {
-      const action = actions.find(item => item.causes);
-      return action.causes[0];
-    }
-  }
-
-  async getAPIData() {
-    const { cli, request, path } = this;
-    const url = this.apiUrl;
-    cli.updateSpinner(`Querying API for ${path}`);
-    return request.json(url);
-  }
-
-  async getConsoleText() {
-    const { cli, request, path } = this;
-    cli.updateSpinner(`Querying console text for ${path}`);
-    const data = await request.text(this.consoleUrl);
-    return data.replace(/\r/g, '');
-  }
-
-  getCacheKey() {
-    return this.path
-      .replace(/job\//, '')
-      .replace(/\//g, '-')
-      .replace(/-$/, '');
-  }
-
-  async parseConsoleText() {
-    let text;
-    try {
-      text = await this.getConsoleText();
-    } catch (err) {
-      this.failures = [
-        new NCUFailure({
-          url: this.consoleUrl, builtOn: this.builtOn
-        }, err.message)
-      ];
-      return this.failures;
-    }
-
-    const parser = new CIFailureParser(this, text);
-    let results = parser.parse();
-    if (!results) {
-      results = [
-        new CIResult({ url: this.jobUrl, builtOn: this.builtOn }, 'Unknown')
-      ];
-    }
-
-    this.failures = results;
-    return results;
-  }
-}
-
-// TODO(joyeecheung): do not cache pending jobs
-const jobCache = new Cache();
-jobCache.wrap(Job, {
-  getConsoleText() {
-    return { key: this.getCacheKey(), ext: '.txt' };
-  },
-  getAPIData() {
-    return { key: this.getCacheKey(), ext: '.json' };
-  }
-});
-
-class TestBuild extends Job {
-  constructor(cli, request, path, tree) {
-    super(cli, request, path, tree);
-
-    // Should be assigned in getResults()
-    this.result = null;
-    this.params = {};
-    this.change = {};
-    this.date = undefined;
-    this.builds = {
-      failed: [], aborted: [], pending: [], unstable: []
-    };
-    this.failures = [];
-    this.builtOn = undefined;
-  }
-
-  setBuildData({ result, changeSet, actions, timestamp, builtOn }) {
-    const params = actions.find(item => typeof item.parameters === 'object');
-    params.parameters.forEach(pair => {
-      this.params[pair.name] = pair.value;
-    });
-
-    this.change = changeSet.items[0] || {};
-    this.date = new Date(timestamp);
-    this.result = result;
-    this.builtOn = builtOn;
-  }
-
-  get sourceURL() {
-    const { params } = this;
-
-    if (params.PR_ID) {  // from a node-test-pull-request build
-      const owner = params.TARGET_GITHUB_ORG;
-      const repo = params.TARGET_REPO_NAME;
-      const prid = params.PR_ID;
-      return `https://github.com/${owner}/${repo}/pull/${prid}/`;
-    }
-
-    if (params.GITHUB_ORG) {  // from a node-test-commit build
-      const owner = params.GITHUB_ORG;
-      const repo = params.REPO_NAME;
-      const prm = params.GIT_REMOTE_REF.match(/refs\/pull\/(\d+)\/head/);
-      if (prm) {
-        return `https://github.com/${owner}/${repo}/pull/${prm[1]}/`;
-      } else {
-        const result =
-          `https://api.github.com/repos/${owner}/${repo}/git/` +
-          params.GIT_REMOTE_REF;
-        return result;
-      }
-    }
-  }
-
-  get commit() {
-    const { change } = this;
-    if (!change.commitId) {
-      return 'Unknown';
-    }
-    return `[${shortSha(change.commitId)}] ${change.msg}`;
-  }
-
-  get author() {
-    const { change } = this;
-    if (!change.author) {
-      return 'Unknown';
-    }
-    return `${change.author.fullName} <${change.authorEmail}>`;
-  }
-
-  displayHeader() {
-    const { cli, result, change } = this;
-    cli.separator('Summary');
-    cli.table('Result', resultName(result));
-    cli.table('URL', this.jobUrl);
-    cli.table('Source', this.sourceURL);
-    cli.table('Commit', this.commit);
-    cli.table('Date', change.date);
-    cli.table('Author', this.author);
-  }
-
-  displayFailure(failure) {
-    const { cli } = this;
-    const { url, reason } = failure;
-    cli.separator(getNodeName(url));
-    cli.table('URL', url);
-    if (failure.type) {
-      cli.table('Type', failure.type);
-    }
-    if (failure.builtOn) {
-      cli.table('Built On', failure.builtOn);
-    }
-    if (!reason.includes('\n') && reason.length < 40) {
-      cli.table('Reason', chalk.red(reason));
-    } else {
-      cli.table('Reason', '');
-      const lines = reason.split('\n');
-      cli.log(`  ${chalk.red(lines[0])}`);
-      for (let i = 1; i < lines.length; ++i) {
-        cli.log(`  ${lines[i]}`);
-      }
-    }
-  }
-
-  displayBuilds() {
-    const { cli, failures, builds } = this;
-    for (const failure of failures) {
-      if (failure !== undefined) {
-        this.displayFailure(failure);
-      }
-    }
-    cli.separator('Other builds');
-    for (const aborted of builds.aborted) {
-      cli.table('Aborted', getUrl(aborted.url));
-    }
-    for (const pending of builds.pending) {
-      cli.table('Pending', getUrl(pending.url));
-    }
-    for (const unstable of builds.unstable) {
-      cli.table('Unstable', getUrl(unstable.url));
-    }
-  }
-
-  display() {
-    this.displayHeader();
-    this.displayBuilds();
-  }
-
-  formatAsMarkdown() {
-    if (this.result === SUCCESS) {
-      return `Job ${this.jobUrl} is green.`;
-    }
-    const { failures } = this;
-    let output = `Failures in job ${this.jobUrl}\n\n`;
-    for (const failure of failures) {
-      if (failure === undefined) continue;
-      output += `#### [${getNodeName(failure.url)}](${failure.url})`;
-      if (!failure.reason.includes('\n') && failure.reason.length < 20) {
-        const builtOn = failure.builtOn ? `On ${failure.builtOn}: ` : '';
-        output += `\n\n${builtOn}${failure.reason}\n`;
-      } else {
-        output += '\n\n';
-        const builtOn = failure.builtOn ? ` on ${failure.builtOn}:` : '';
-        output += fold(`See failures${builtOn}`, failure.reason) + '\n\n';
-      }
-    }
-    return output;
-  }
-
-  formatAsJson() {
-    const { jobUrl, failures, sourceURL } = this;
-
-    const result = failures.map(item => Object.assign({
-      source: sourceURL,
-      upstream: jobUrl
-    }, item));
-
-    return JSON.parse(JSON.stringify(result));
-  }
 }
 
 function getHighlight(f) {
@@ -392,10 +78,6 @@ function markdownRow(...args) {
 
 function getMachineUrl(name) {
   return `[${name}](https://${CI_DOMAIN}/computer/${name}/)`;
-}
-
-function pad(any, length) {
-  return (any + '').padEnd(length);
 }
 
 const kHealthKeys = ['success', 'pending', 'aborted', 'failed', 'unstable'];
@@ -597,17 +279,25 @@ class CommitBuild extends TestBuild {
   }
 
   getBuilds({ result, subBuilds }) {
-    if (result === SUCCESS) {
+    if (result === statusType.SUCCESS) {
       const builds = this.builds = {
         failed: [], aborted: [], pending: [], unstable: []
       };
-      return { result: SUCCESS, builds };
+      return { result: statusType.SUCCESS, builds };
     }
 
-    const failed = subBuilds.filter(build => build.result === FAILURE);
-    const aborted = subBuilds.filter(build => build.result === ABORTED);
-    const pending = subBuilds.filter(build => build.result === null);
-    const unstable = subBuilds.filter(build => build.result === UNSTABLE);
+    const failed = subBuilds.filter(build => {
+      return build.result === statusType.FAILURE;
+    });
+    const aborted = subBuilds.filter(build => {
+      return build.result === statusType.ABORTED;
+    });
+    const pending = subBuilds.filter(build => {
+      return build.result === null;
+    });
+    const unstable = subBuilds.filter(build => {
+      return build.result === statusType.UNSTABLE;
+    });
 
     // build: { buildNumber, jobName, result, url }
     const builds = this.builds = { failed, aborted, pending, unstable };
@@ -630,7 +320,7 @@ class CommitBuild extends TestBuild {
     }
     this.setBuildData(data);
     // No builds at all
-    if (data.result === FAILURE && !data.subBuilds.length) {
+    if (data.result === statusType.FAILURE && !data.subBuilds.length) {
       const failure = new BuildFailure(this, 'Failed to trigger sub builds');
       this.failures = [failure];
       return {
@@ -642,7 +332,7 @@ class CommitBuild extends TestBuild {
 
     const { result, builds } = this.getBuilds(data);
 
-    if (result !== FAILURE) {
+    if (result !== statusType.FAILURE) {
       return { result, builds, failures: [] };
     }
 
@@ -759,164 +449,6 @@ class PRBuild extends TestBuild {
   }
 }
 
-class CITGMBuild extends TestBuild {
-  constructor(cli, request, job) {
-    const { jobid, noBuild } = job;
-    const path = noBuild
-      ? `job/citgm-smoker-nobuild/${jobid}/`
-      : `job/citgm-smoker/${jobid}/`;
-
-    const tree = CITGM_MAIN_TREE;
-
-    super(cli, request, path, tree);
-
-    this.id = jobid;
-    this.noBuild = noBuild;
-  }
-
-  async getResults() {
-    let headerData;
-    try {
-      headerData = await this.getBuildData('Summary');
-    } catch (err) {
-      this.failures = [
-        new NCUFailure({ url: this.apiUrl }, err.message)
-      ];
-      return this.failures;
-    }
-    const { result } = headerData;
-
-    this.setBuildData(headerData);
-
-    // CITGM jobs store results in a different location than
-    // they do summary data, so we need to update the endpoint
-    // and issue a second API call in order to fetch result data.
-    this.tree = CITGM_REPORT_TREE;
-    this.updatePath(true);
-
-    let resultData;
-    try {
-      resultData = await this.getBuildData('Results');
-    } catch (err) {
-      this.failures = [
-        new NCUFailure({ url: this.apiUrl }, err.message)
-      ];
-      return this.failures;
-    }
-
-    this.results = this.parseResults(resultData);
-
-    // Update id again so that it correctly displays in Summary output.
-    this.updatePath(false);
-
-    return { result };
-  }
-
-  parseResults(data) {
-    const { childReports, totalCount, skipCount, failCount } = data;
-    const results = { all: {}, failures: {}, statistics: {} };
-
-    const passCount = totalCount - failCount - skipCount;
-    results.statistics.passed = passCount;
-    results.statistics.total = totalCount;
-    results.statistics.skipped = skipCount;
-    results.statistics.failed = failCount;
-
-    childReports.forEach(platform => {
-      const cases = flatten(platform.result.suites[0].cases);
-      const url = platform.child.url;
-      const nodeName = getNodeName(url);
-
-      results.all[nodeName] = { url, modules: cases };
-
-      const failedModules = cases.filter(c => c.status === 'FAILED');
-      results.failures[nodeName] = { url, modules: failedModules };
-    });
-
-    return results;
-  }
-
-  updatePath(testReport) {
-    const { id, noBuild } = this;
-    if (testReport) {
-      this.path = noBuild
-        ? `job/citgm-smoker-nobuild/${id}/testReport/`
-        : `job/citgm-smoker/${id}/testReport/`;
-    } else {
-      this.path = noBuild
-        ? `job/citgm-smoker-nobuild/${id}/`
-        : `job/citgm-smoker/${id}/`;
-    }
-  }
-
-  displayBuilds() {
-    const { cli, results } = this;
-    const { failed, skipped, passed, total } = results.statistics;
-
-    cli.separator('Statistics');
-    console.table({
-      Failed: failed,
-      Skipped: skipped,
-      Passed: passed,
-      Total: total
-    });
-
-    cli.separator('Failures');
-    const output = {};
-    for (const platform in results.failures) {
-      const modules = results.failures[platform].modules;
-      const failures = modules.map(f => f.name);
-
-      output[platform] = failures;
-    }
-
-    console.table(output);
-  }
-
-  formatAsJson() {
-    const { jobUrl, results, sourceURL } = this;
-
-    const result = {
-      source: sourceURL,
-      upstream: jobUrl,
-      ...results.statistics,
-      ...results.failures
-    };
-
-    return JSON.parse(JSON.stringify(result));
-  }
-
-  formatAsMarkdown() {
-    const { jobUrl, result, results, id } = this;
-
-    let output = `# CITGM Data for [${id}](${jobUrl})\n\n`;
-
-    const { failed, skipped, passed, total } = results.statistics;
-
-    output += `## Statistics for job [${id}](${jobUrl})\n\n`;
-    output += '|  FAILED  |  SKIPPED  |  PASSED  |  TOTAL  |\n';
-    output += '| -------- | --------- | -------- | ------- |\n';
-    output += `| ${pad(failed, 8)} | ${pad(skipped, 9)} |`;
-    output += ` ${pad(passed, 8)} | ${pad(total, 7)} |\n\n`;
-
-    if (result === SUCCESS) {
-      output += `Job [${id}](${jobUrl}) is green.`;
-      return output;
-    }
-
-    output += `## Failures in job [${id}](${jobUrl})\n\n`;
-    for (const failure in results.failures) {
-      const data = results.failures[failure];
-      output += `### [${failure}](${data.url})\n\n`;
-
-      const failures = data.modules.map(f => `* ${f.name}`);
-      const items = failures.length > 0 ? `${failures.join('\n')}` : 'None.';
-      output += `${items}\n\n`;
-    }
-    return output;
-  }
-}
-
 class DailyBuild extends TestBuild {
   constructor(cli, request, id) {
     const path = `job/node-daily-master/${id}/`;
@@ -955,11 +487,11 @@ async function listBuilds(cli, request, type) {
 
   const result = await request.json(url);
   const builds = result.builds;
-  const failed = filterBuild(builds, FAILURE);
-  const aborted = filterBuild(builds, ABORTED);
+  const failed = filterBuild(builds, statusType.FAILURE);
+  const aborted = filterBuild(builds, statusType.ABORTED);
   const pending = filterBuild(builds, null);
-  const unstable = filterBuild(builds, UNSTABLE);
-  const success = filterBuild(builds, SUCCESS);
+  const unstable = filterBuild(builds, statusType.UNSTABLE);
+  const success = filterBuild(builds, statusType.SUCCESS);
   cli.stopSpinner('Done');
 
   return {
@@ -1006,7 +538,9 @@ class FannedBuild extends Job {
       return this.failures;
     }
 
-    const failedPhase = data.subBuilds.find(build => build.result === FAILURE);
+    const failedPhase = data.subBuilds.find(build => {
+      return build.result === statusType.FAILURE;
+    });
 
     if (!failedPhase) {
       return this.parseConsoleText();
@@ -1088,7 +622,7 @@ class NormalBuild extends Job {
     const { result, runs, builtOn } = data;
     this.builtOn = builtOn;
 
-    if (result !== FAILURE) {
+    if (result !== statusType.FAILURE) {
       this.failures = [];
       return this.failures;
     }
@@ -1104,7 +638,9 @@ class NormalBuild extends Job {
       return this.failures;
     }
 
-    const failed = runs.filter(run => run.result === FAILURE);
+    const failed = runs.filter(run => {
+      return run.result === statusType.FAILURE;
+    });
 
     if (!failed.length) {
       return this.parseConsoleText();
